moved: Paging-flow helpers from ui to domain module
